+++ LEColorPicker/UIColor+YUVSpace.m
@@ -20,7 +20,7 @@
     float y = 0.0;
     
     if (color) {
-        int numComponents = CGColorGetNumberOfComponents(color.CGColor);
+        size_t numComponents = CGColorGetNumberOfComponents(color.CGColor);
         
         if (numComponents == 4) {
             const CGFloat *components = CGColorGetComponents(color.CGColor);
@@ -46,7 +46,7 @@
     float u = 0.0;
     
     if (color) {
-        int numComponents = CGColorGetNumberOfComponents(color.CGColor);
+        size_t numComponents = CGColorGetNumberOfComponents(color.CGColor);
         
         if (numComponents == 4) {
             const CGFloat *components = CGColorGetComponents(color.CGColor);
@@ -71,7 +71,7 @@
     float v = 0.0;
     
     if (color) {
-        int numComponents = CGColorGetNumberOfComponents(color.CGColor);
+        size_t numComponents = CGColorGetNumberOfComponents(color.CGColor);
         
         if (numComponents == 4) {
             const CGFloat *components = CGColorGetComponents(color.CGColor);
